Respect adapter-level options in adapter.receive()
According to [Skipper's docs](https://github.com/balderdashy/skipper#specifying-options), the adapter factory method may take an `options` object that will become the defaults for all subsequent operations on that adapter, unless overriden in subsequent calls to `adapter.receive()` or `file.upload()`.

This behaviour was not respected - the adapter-level defaults were not applied. This PR remedies that.